third_party/blink/web_tests/css-parser/appearance-warning-slider-vertical-usecounter.html

<!DOCTYPE html>
<link rel=author href="mailto:dizhangg@chromium.org">
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>

<button id="button">foo</button>

<script>
const button = document.getElementById('button');
const CSSValueAppearanceSliderVertical = 4556;

test(() => {
  assert_false(internals.isUseCounted(document, CSSValueAppearanceSliderVertical),
    'Having an element with no appearance should not activate the use counter.');
  button.style = `appearance: auto`;
  assert_false(internals.isUseCounted(document, CSSValueAppearanceSliderVertical),
    'Having an element with appearance value "auto" should not activate the use counter.');

  button.style = `appearance: slider-vertical`;
  assert_true(internals.isUseCounted(document, CSSValueAppearanceSliderVertical),
    'Having an element with to be deprecated appearance value "slider-vertical" should activate the use counter.');

  button.removeAttribute('appearance');
  internals.clearUseCounter(document, CSSValueAppearanceSliderVertical);
  assert_false(internals.isUseCounted(document, CSSValueAppearanceSliderVertical));

  button.style = `-webkit-appearance: auto`;
  assert_false(internals.isUseCounted(document, CSSValueAppearanceSliderVertical),
    'Having an element with -webkit-appearance value "auto" should not activate the use counter.');

  button.style = `-webkit-appearance: slider-vertical`;
  assert_true(internals.isUseCounted(document, CSSValueAppearanceSliderVertical),
    'Having an element with to be deprecated -webkit-appearance value "slider-vertical" should activate the use counter.');
});
</script>
